refactor: MainActivity에 MVP 패턴 적용
