+++ app/src/androidTest/java/woowacourse/movie/MainActivityTest.kt
@@ -8,6 +8,7 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import org.junit.Rule
 import org.junit.Test
 import org.junit.jupiter.api.DisplayName
+import woowacourse.movie.main.MainActivity
 
 class MainActivityTest {
     @get:Rule
