+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,13 @@
         tools:targetApi="31">
         <activity
             android:name=".ReserveActivity"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".main.MainActivity"
             android:exported="false" />
         <activity
             android:name=".ReservationResultActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
