+++ app/src/main/java/woowacourse/movie/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.main
 
 import android.os.Bundle
 import android.widget.ListView
@@ -6,10 +6,14 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import woowacourse.movie.MoviesAdapter
+import woowacourse.movie.R
+import woowacourse.movie.ReserveActivity
 import woowacourse.movie.domain.Movie
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity(), MainContract.View {
     private val moviesView: ListView by lazy { findViewById(R.id.lv_movies) }
+    private val presenter: MainPresenter = MainPresenter(this)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -20,12 +24,11 @@ class MainActivity : AppCompatActivity() {
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
-        val movies: List<Movie> = DefaultMovies.movies
 
-        initAdapter(movies)
+        presenter.initMovies()
     }
 
-    private fun initAdapter(movies: List<Movie>) {
+    override fun showMovies(movies: List<Movie>) {
         moviesView.adapter =
             MoviesAdapter(movies) { movie ->
                 val intent = ReserveActivity.newIntent(this, movie)
