+++ app/src/main/java/woowacourse/movie/main/MainContract.kt
@@ -1,0 +1,13 @@
+package woowacourse.movie.main
+
+import woowacourse.movie.domain.Movie
+
+interface MainContract {
+    interface View {
+        fun showMovies(movies: List<Movie>)
+    }
+
+    interface Presenter {
+        fun initMovies()
+    }
+}
